+++ core/src/com/game/planetdefense/Actors/Asteroid.java
@@ -6,6 +6,7 @@ import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.MathUtils;
+import com.badlogic.gdx.math.Polygon;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.scenes.scene2d.Actor;
@@ -16,6 +17,7 @@ public class Asteroid extends Actor implements Pool.Poolable{
     private Animation<TextureRegion> animation;
     private Sprite sprite;
     private Rectangle position;
+    private Polygon collision;
     private Vector2 target;
     private float state_time;
 
@@ -30,6 +32,7 @@ public class Asteroid extends Actor implements Pool.Poolable{
         this.sprite = new Sprite();
         this.sprite.setBounds( position.getX(), position.getY(), position.getWidth(), position.getHeight());
         this.target = new Vector2(0,0);
+        this.collision = new Polygon();
     }
 
     @Override
@@ -69,6 +72,7 @@ public class Asteroid extends Actor implements Pool.Poolable{
         Vector2 temp = new Vector2(0,0);
         float angle = temp.set(target.x,target.y).sub(position.getX() + sprite.getOriginX(), position.getY() + sprite.getOriginY()).angle();
         this.sprite.setRotation(angle);
+        this.collision.setRotation(angle);
     }
 
     private void moveToTarget(float delta){
@@ -78,13 +82,17 @@ public class Asteroid extends Actor implements Pool.Poolable{
         position.setPosition(help);*/
         this.position.x += speed * delta * MathUtils.cos((float)((Math.PI / 180) * ( sprite.getRotation())));
         this.position.y += speed * delta * MathUtils.sin((float)((Math.PI / 180) * ( sprite.getRotation())));
+        this.collision.setPosition(this.position.getX(), this.position.getY());
     }
 
     public void setAsteroid(float x, float y, float width, float height){
-        this.position.set(x - width/2, y - height/2, width, height);
+        this.position.set(x /*- width/2*/, y /*- height/2*/, width, height);
         this.sprite.setBounds(this.position.getX(), this.position.getY(), this.position.getWidth(), this.position.getHeight());
         this.sprite.setOriginCenter();
-        this.position.setSize(this.position.getWidth() - (this.position.getWidth() * 0.2f), this.position.getHeight() - (this.position.getHeight() * 0.2f));
+        //this.position.setSize(this.position.getWidth() - (this.position.getWidth() * 0.2f), this.position.getHeight() - (this.position.getHeight() * 0.2f));
+        this.collision.setVertices(new float[]{0,0,position.width * 0.99f,0,position.width * 0.99f,position.height * 0.99f,0,position.height * 0.99f});
+        this.collision.setPosition(this.position.getX(), this.position.getY());
+        this.collision.setOrigin(this.position.getWidth()/2, this.position.getHeight()/2);
         Gdx.app.log("Asteroid position", " " + sprite.getX() + " " + sprite.getY());
     }
 
@@ -140,40 +148,6 @@ public class Asteroid extends Actor implements Pool.Poolable{
     public Rectangle getRectangle(){
         return position;
     }
+    public Polygon getPolygon(){return collision;}
 
-    /* Rectangle bounds;
-   Polygon polygon;
-
-   Rectangle bounds2;
-   Polygon polygon2;
-...
-   @Override
-   public void create() {
-...
-      bounds = new Rectangle(0, 0, 32, 20);
-      polygon = new Polygon(new float[]{0,0,bounds.width,0,bounds.width,bounds.height,0,bounds.height});
-      polygon.setOrigin(bounds.width/2, bounds.height/2);
-
-      bounds2 = new Rectangle(0, 0, 32, 20);
-      polygon2 = new Polygon(new float[]{0,0,bounds2.width,0,bounds2.width,bounds2.height,0,bounds2.height});
-      polygon2.setOrigin(bounds2.width/2, bounds2.height/2);
-
-...
-}
-   @Override
-   public void render() {
-...
-
-
-      polygon.setPosition(car1.x, car1.y);
-      polygon.setRotation(car1.rotation);
-      polygon2.setPosition(car2.x, car2.y);
-      polygon2.setRotation(car2.rotation);
-
-...
-        if(Intersector.overlapConvexPolygons(polygon, polygon2)){
-            //COLLISION DON'T HAPPEN!!!
-        }
-...
-}*/
 }
