+++ core/src/com/game/planetdefense/Actors/Missile.java
@@ -5,6 +5,7 @@ import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.MathUtils;
+import com.badlogic.gdx.math.Polygon;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.scenes.scene2d.Actor;
@@ -18,6 +19,7 @@ public class Missile extends Actor implements Pool.Poolable {
     private LaserType laserType;
     private Sprite sprite;
     private Rectangle position;
+    private Polygon collision;
     private Vector2 target;
     private float damage;
     public float flight_time;
@@ -31,6 +33,7 @@ public class Missile extends Actor implements Pool.Poolable {
         this.sprite.setBounds( position.getX(), position.getY(), position.getWidth(), position.getHeight());
         this.target = new Vector2(0,0);
         flight_time = 0;
+        collision = new Polygon();
     }
 
     @Override
@@ -72,6 +75,7 @@ public class Missile extends Actor implements Pool.Poolable {
         Vector2 temp = new Vector2(0,0);
         float angle = temp.set(target.x,target.y).sub(position.getX() + sprite.getOriginX(), position.getY() + sprite.getOriginY()).angle();
         this.sprite.setRotation(angle - 90f);
+        this.collision.setRotation(angle - 90f);
     }
 
     private void moveToTarget(float delta){
@@ -81,6 +85,7 @@ public class Missile extends Actor implements Pool.Poolable {
         position.setPosition(help);*/
         this.position.x += laserType.getSpeed() * delta * MathUtils.cos((float)((Math.PI / 180) * ( sprite.getRotation() + 90)));
         this.position.y += laserType.getSpeed() * delta * MathUtils.sin((float)((Math.PI / 180) * ( sprite.getRotation() + 90)));
+        this.collision.setPosition(this.position.getX(), this.position.getY());
     }
 
     public void moveForward(float distance){
@@ -94,6 +99,9 @@ public class Missile extends Actor implements Pool.Poolable {
         this.position.set(x, y, laserType.getWidth(), laserType.getHeight());
         this.sprite.setBounds(this.position.getX(), this.position.getY(), this.position.getWidth(), this.position.getHeight());
         this.sprite.setOrigin(this.sprite.getWidth()/2, 0);
+        this.collision.setVertices(new float[]{0,0,position.width,0,position.width,position.height,0,position.height});
+        this.collision.setPosition(this.position.getX(), this.position.getY());
+        this.collision.setOrigin(this.position.getWidth()/2, 0);
         rotateToTarget();
         this.damage = 5f + (UpgradeType.DmgBonus.getUpgradeLvl() * UpgradeType.DmgBonus.getUpgradeValue());
         Gdx.app.log("DMG", ""+damage);
@@ -125,4 +133,41 @@ public class Missile extends Actor implements Pool.Poolable {
     public Rectangle getRectangle(){
         return position;
     }
+    public Polygon getPolygon(){return collision;}
+
+        /* Rectangle bounds;
+   Polygon polygon;
+
+   Rectangle bounds2;
+   Polygon polygon2;
+...
+   @Override
+   public void create() {
+...
+      bounds = new Rectangle(0, 0, 32, 20);
+      polygon = new Polygon(new float[]{0,0,bounds.width,0,bounds.width,bounds.height,0,bounds.height});
+      polygon.setOrigin(bounds.width/2, bounds.height/2);
+
+      bounds2 = new Rectangle(0, 0, 32, 20);
+      polygon2 = new Polygon(new float[]{0,0,bounds2.width,0,bounds2.width,bounds2.height,0,bounds2.height});
+      polygon2.setOrigin(bounds2.width/2, bounds2.height/2);
+
+...
+}
+   @Override
+   public void render() {
+...
+
+
+      polygon.setPosition(car1.x, car1.y);
+      polygon.setRotation(car1.rotation);
+      polygon2.setPosition(car2.x, car2.y);
+      polygon2.setRotation(car2.rotation);
+
+...
+        if(Intersector.overlapConvexPolygons(polygon, polygon2)){
+            //COLLISION DON'T HAPPEN!!!
+        }
+...
+}*/
 }
