+++ core/src/com/game/planetdefense/GameStage.java
@@ -2,6 +2,7 @@ package com.game.planetdefense;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
+import com.badlogic.gdx.math.Intersector;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.Stage;
@@ -140,7 +141,7 @@ public class GameStage extends Stage {
             while(missile_iterator.hasNext()){
                 Missile missile = missile_iterator.next();
                 //collision missile with asteroid
-                if(asteroid.getRectangle().overlaps(missile.getRectangle())){
+                if(Intersector.overlapConvexPolygons(asteroid.getPolygon(), missile.getPolygon())){
                     asteroid.setHp(asteroid.getHp() - missile.getDamage());
                     if(asteroid.getHp() > 0){
                         missile_pool.free(missile);
